split plot into modules
